Create Figure_8.R
This file will generate figure 8 from the paper (anaysis of data from Demas et al 2009 and Wong et al 1993). It assumes that the c code in this repository has been compiled and that the necessary data files have been downloaded and are in the code's directory. The data is freely available at https://portal.carmen.org.uk/.